+++ sdk/botservice/azure-resourcemanager-botservice/src/main/java/com/azure/resourcemanager/botservice/implementation/AzureBotServiceImpl.java
@@ -44,8 +44,6 @@ import reactor.core.publisher.Mono;
 /** Initializes a new instance of the AzureBotServiceImpl type. */
 @ServiceClient(builder = AzureBotServiceBuilder.class)
 public final class AzureBotServiceImpl implements AzureBotService {
-    private final ClientLogger logger = new ClientLogger(AzureBotServiceImpl.class);
-
     /** Azure Subscription ID. */
     private final String subscriptionId;
 
@@ -343,7 +341,7 @@ public final class AzureBotServiceImpl implements AzureBotService {
                             managementError = null;
                         }
                     } catch (IOException | RuntimeException ioe) {
-                        logger.logThrowableAsWarning(ioe);
+                        LOGGER.logThrowableAsWarning(ioe);
                     }
                 }
             } else {
@@ -402,4 +400,6 @@ public final class AzureBotServiceImpl implements AzureBotService {
             return Mono.just(new String(responseBody, charset));
         }
     }
+
+    private static final ClientLogger LOGGER = new ClientLogger(AzureBotServiceImpl.class);
 }