+++ sdk/communication/azure-communication-chat/src/main/java/com/azure/communication/chat/implementation/ChatThreadsImpl.java
@@ -72,7 +72,7 @@ public final class ChatThreadsImpl {
      */
     @Host("{endpoint}")
     @ServiceInterface(name = "AzureCommunicationCh")
-    private interface ChatThreadsService {
+    public interface ChatThreadsService {
         @Get("/chat/threads/{chatThreadId}/readReceipts")
         @ExpectedResponses({200})
         @UnexpectedResponseExceptionType(CommunicationErrorResponseException.class)