+++ sdk/recoveryservicesbackup/azure-resourcemanager-recoveryservicesbackup/src/main/java/com/azure/resourcemanager/recoveryservicesbackup/models/AzureWorkloadSqlRecoveryPointExtendedInfo.java
@@ -5,8 +5,6 @@
 package com.azure.resourcemanager.recoveryservicesbackup.models;
 
 import com.azure.core.annotation.Fluent;
-import com.azure.core.util.logging.ClientLogger;
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.OffsetDateTime;
 import java.util.List;
@@ -14,8 +12,6 @@ import java.util.List;
 /** Extended info class details. */
 @Fluent
 public final class AzureWorkloadSqlRecoveryPointExtendedInfo {
-    @JsonIgnore private final ClientLogger logger = new ClientLogger(AzureWorkloadSqlRecoveryPointExtendedInfo.class);
-
     /*
      * UTC time at which data directory info was captured
      */