+++ sdk/resourcemanager/azure-resourcemanager-appplatform/src/main/java/com/azure/resourcemanager/appplatform/implementation/SpringAppImpl.java
@@ -18,6 +18,7 @@ import com.azure.resourcemanager.appplatform.models.SpringAppDeployment;
 import com.azure.resourcemanager.appplatform.models.SpringAppDeployments;
 import com.azure.resourcemanager.appplatform.models.SpringAppDomains;
 import com.azure.resourcemanager.appplatform.models.SpringAppServiceBindings;
+import com.azure.resourcemanager.appplatform.models.SpringConfigurationService;
 import com.azure.resourcemanager.appplatform.models.SpringService;
 import com.azure.resourcemanager.appplatform.models.TemporaryDisk;
 import com.azure.resourcemanager.appplatform.models.UserSourceType;
@@ -27,6 +28,9 @@ import com.azure.resourcemanager.resources.fluentcore.model.Creatable;
 import reactor.core.publisher.Mono;
 
 import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
 
 public class SpringAppImpl
@@ -143,6 +147,23 @@ public class SpringAppImpl
         if (innerModel().properties() == null) {
             innerModel().withProperties(new AppResourceProperties());
         }
+    }
+
+    @Override
+    public boolean hasConfigurationServiceBinding() {
+        Map<String, Map<String, Object>> addonConfigs = innerModel().properties().addonConfigs();
+        if (addonConfigs == null) {
+            return false;
+        }
+        SpringConfigurationService configurationService = parent().getDefaultConfigurationService();
+        if (configurationService == null) {
+            return false;
+        }
+        return addonConfigs.get(Constants.APPLICATION_CONFIGURATION_SERVICE_KEY) != null
+            && Objects.equals(
+            addonConfigs.get(Constants.APPLICATION_CONFIGURATION_SERVICE_KEY).get(Constants.BINDING_RESOURCE_ID),
+            configurationService.id()
+        );
     }
 
     @Override
@@ -305,6 +326,36 @@ public class SpringAppImpl
     SpringAppImpl addActiveDeployment(SpringAppDeploymentImpl deployment) {
         withActiveDeployment(deployment.name());
         springAppDeploymentToCreate = deployment;
+        return this;
+    }
+
+    @Override
+    public SpringAppImpl withConfigurationServiceBinding() {
+        ensureProperty();
+        Map<String, Map<String, Object>> addonConfigs = innerModel().properties().addonConfigs();
+        if (addonConfigs == null) {
+            addonConfigs = new HashMap<>();
+            innerModel().properties().withAddonConfigs(addonConfigs);
+        }
+        Map<String, Object> configurationServiceConfigs = addonConfigs.computeIfAbsent(Constants.APPLICATION_CONFIGURATION_SERVICE_KEY, k -> new HashMap<>());
+        configurationServiceConfigs.put(Constants.BINDING_RESOURCE_ID, parent().getDefaultConfigurationService().id());
+        return this;
+    }
+
+    @Override
+    public SpringAppImpl withoutConfigurationServiceBinding() {
+        if (innerModel().properties() == null) {
+            return this;
+        }
+        Map<String, Map<String, Object>> addonConfigs = innerModel().properties().addonConfigs();
+        if (addonConfigs == null) {
+            return this;
+        }
+        Map<String, Object> configurationServiceConfigs = addonConfigs.get(Constants.APPLICATION_CONFIGURATION_SERVICE_KEY);
+        if (configurationServiceConfigs == null) {
+            return this;
+        }
+        configurationServiceConfigs.put(Constants.BINDING_RESOURCE_ID, "");
         return this;
     }
 }