+++ sdk/resourcemanager/azure-resourcemanager-appplatform/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## 2.15.0-beta.1 (Unreleased)
 
+### Features Added
+
+- Supported Enterprise Tier `Configuration Service`.
+- Supported Enterprise Tier `Build Service`.
+- Supported Enterprise Tier binding `Spring App` to `Configuration Service`.
+- Supported Enterprise Tier `Spring App Deployment` with `Jar` and `Maven Source Code`.
+- Supported `Java_17` runtime for all tiers.
+- Supported `jvmOptions()` in `SpringAppDeployment` for all tiers.
+- Supported Enterprise Tier `runtimeVersion()` in `SpringAppDeployment`.
+
 ### Breaking Changes
 
 - Removed `createdTime` from `AppResourceProperties` and `DeploymentResourceProperties`.